finish opencv impl

+++ src/csv/OutputCSVData.cpp
@@ -4,7 +4,7 @@
 
 namespace ImageComparison
 {
-    OutputCSVData::OutputCSVData(InputCSVData imagePair, uint64_t simularityScore, float secondsTaken) noexcept
+    OutputCSVData::OutputCSVData(InputCSVData imagePair, double simularityScore, float secondsTaken) noexcept
     : mImagePair(std::move(imagePair))
     , mSimularityScore(simularityScore)
     , mSecondsTaken(secondsTaken) {}
@@ -12,13 +12,13 @@ namespace ImageComparison
     OutputCSVData::OutputCSVData(OutputCSVData&& other) noexcept
     : mImagePair(std::move(other.mImagePair))
     , mSimularityScore(std::exchange(other.mSimularityScore, 0))
-    , mSecondsTaken(std::exchange(other.mSecondsTaken, 0.0)) {}
+    , mSecondsTaken(std::exchange(other.mSecondsTaken, 0.0f)) {}
 
     OutputCSVData& OutputCSVData::operator=(OutputCSVData&& other) noexcept {
         if (this != &other) {
             mImagePair = std::move(other.mImagePair);
-            mSimularityScore = std::exchange(other.mSimularityScore, 0);
-            mSecondsTaken = std::exchange(other.mSecondsTaken, 0.0);
+            mSimularityScore = std::exchange(other.mSimularityScore, 0.0);
+            mSecondsTaken = std::exchange(other.mSecondsTaken, 0.0f);
         }
 
         return *this;
@@ -28,7 +28,7 @@ namespace ImageComparison
         return mImagePair;
     }
 
-    uint64_t OutputCSVData::GetSimularityScore() const noexcept {
+    double OutputCSVData::GetSimularityScore() const noexcept {
         return mSimularityScore;
     }
